move commands into individual modules
