+++ setup.py
@@ -56,17 +56,17 @@ setup(
     entry_points={
         'console_scripts': ['lymph = lymph.cli.main:main'],
         'lymph.cli': [
+            'discover = lymph.cli.discover:DiscoverCommand',
+            'emit = lymph.cli.emit:EmitCommand',
             'help = lymph.cli.help:HelpCommand',
-            'list = lymph.cli.base:ListCommand',
-            'tail = lymph.cli.tail:TailCommand',
+            'inspect = lymph.cli.inspect:InspectCommand',
             'instance = lymph.cli.service:InstanceCommand',
+            'list = lymph.cli.list:ListCommand',
             'node = lymph.cli.service:NodeCommand',
             'request = lymph.cli.request:RequestCommand',
-            'discover = lymph.cli.request:DiscoverCommand',
-            'inspect = lymph.cli.request:InspectCommand',
-            'subscribe = lymph.cli.request:SubscribeCommand',
-            'emit = lymph.cli.request:EmitCommand',
             'shell = lymph.cli.shell:ShellCommand',
+            'subscribe = lymph.cli.subscribe:SubscribeCommand',
+            'tail = lymph.cli.tail:TailCommand',
         ],
         'nose.plugins.0.10': ['lymph = lymph.testing.nose:LymphPlugin'],
         'pytest11': ['lymph = lymph.testing.pytest'],
